backend/app/models/user.py
from pydantic import BaseModel, Field, EmailStr
from typing import Optional
from bson import ObjectId
from datetime import datetime

class UserBase(BaseModel):
    email: EmailStr

class UserCreate(UserBase):
    password: str

class UserInDB(UserBase):
    id: Optional[str] = Field(alias='_id')
    hashed_password: str
    createdAt: datetime = Field(default_factory=datetime.utcnow)

    class Config:
        populate_by_name = True
        json_encoders = {ObjectId: str}

class User(UserBase):
    id: str = Field(alias='_id')
    createdAt: datetime

    class Config:
        populate_by_name = True
        json_encoders = {ObjectId: str}